requirements.txt
flask>=3.1.0
gdown>=5.2.0
numpy>=2.2.4
opencv-python>=4.11.0.86
pillow>=11.2.1
streamlit>=1.44.1
torch>=2.6.0
torchvision>=0.21.0
